Push up in progress fix picking best child
We found that the best score was not being picked from the frontier.

This works as intended:

intro-to-ec.2048.search=>
(pm/priority-map [4 4 2 0 8 8 0 0 16 16 0 0 32 32 0 0] -854 [4 2 2 4 8 0 4 8 16 0 8 16 0 0 16 32] -480 [2 0 4 4 0 2 4 8 0 0 16 16 0 0 32 32] -720)

{[4 4 2 0 8 8 0 0 16 16 0 0 32 32 0 0] -854,
 [2 0 4 4 0 2 4 8 0 0 16 16 0 0 32 32] -720,
 [4 2 2 4 8 0 4 8 16 0 8 16 0 0 16 32] -480}

However, as you can see by one output from our actual output from the algorithm:

{([[2 0 4 4 0 0 8 8 0 0 16 16 4 0 32 32] -756]
[[4 2 4 4 0 0 8 8 0 0 16 16 0 0 32 32] -756]
[[2 0 2 8 0 0 0 16 0 0 0 32 0 0 2 64] -1134]...)}

The scores are not being put in order. We need a priority map to be applied somewhere.

Provided that can be done in the right place (surely it will be obvious in retrospect),
we should have the solutuion much sooner. Then we can work on a more optimized
heuristic, one that considers more than just 0's

Co-authored-by: UtkarshKr007 <kumar375@morris.umn.edu>